+++ persistence-commons/persistence-commons-protobuf/src/main/java/org/kie/kogito/persistence/protobuf/Constants.java
@@ -1,0 +1,5 @@
+package org.kie.kogito.persistence.protobuf;
+
+public class Constants {
+
+}
