+++ persistence-commons/persistence-commons-protobuf/src/main/java/org/kie/kogito/persistence/protobuf/ProtobufFileMonitorException.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.kie.kogito.index.protobuf;
+package org.kie.kogito.persistence.protobuf;
 
 class ProtobufFileMonitorException extends RuntimeException {
