+++ persistence-commons/persistence-commons-protobuf/src/test/java/org/kie/kogito/persistence/protobuf/ProtostreamProducerTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.kie.kogito.index.protobuf;
+package org.kie.kogito.persistence.protobuf;
 
 import java.io.IOException;
 
@@ -40,8 +40,8 @@ class ProtostreamProducerTest {
         try {
             FileDescriptorSource fileDescriptorSource = protostreamProducer.kogitoTypesDescriptor();
 
-            assertTrue(fileDescriptorSource.getFileDescriptors().containsKey(KOGITO_INDEX_PROTO));
-            assertTrue(fileDescriptorSource.getFileDescriptors().containsKey(KOGITO_TYPES_PROTO));
+            assertTrue(fileDescriptorSource.getFileDescriptors().containsKey(ProtostreamProducer.KOGITO_INDEX_PROTO));
+            assertTrue(fileDescriptorSource.getFileDescriptors().containsKey(ProtostreamProducer.KOGITO_TYPES_PROTO));
         } catch (IOException e) {
             fail("Failed with IOException", e);
         }
