+++ persistence-commons/persistence-commons-protobuf/src/test/java/org/kie/kogito/persistence/protobuf/TestUtils.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.kie.kogito.index.protobuf;
+package org.kie.kogito.persistence.protobuf;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
